Add a concept of ad-hoc rollup rewrite rules
And implement it for comparing a possibly-truncated timestamp to a
pre-truncated literal.  This implements it for comparison operators
but _not_ for BETWEEN (even though the old-rollup matcher _does_ do
BETWEEN) because `x BETWEEN y AND z` is `y <= x AND x <= z`, and that
upper bound can't correctly be rewritten in this way)

+++ soql-analyzer/src/test/scala/com/socrata/soql/analyzer2/TestType.scala
@@ -99,7 +99,7 @@ object TestBoolean extends TestType("boolean", isOrdered = true) {
   def apply(b: Boolean) = if(b) canonicalTrue else canonicalFalse
 }
 
-case class TestNumber(num: Int) extends TestValue {
+case class TestNumber(num: Long) extends TestValue {
   def typ = TestNumber
   def doc = Doc(num)
 }
